photo-server/requirements.txt
PyXDG==0.28
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
jinja2==3.1.2
